complete manual intervention and fix errorCode query for txn validation
